requirements.txt
langchain>=0.1.0
openai>=1.0.0
supabase>=2.0.0
pgvector>=0.2.1
python-dotenv>=1.0.0
tiktoken>=0.5.1
psycopg2-binary>=2.9.0